Add files via upload
Fixed some documentation. Changed the name of test.cpp to main.cpp.
Changed the function name lexeme() to lexer(). Updated the DFSM of id and keyword to be case-insensitive.
The program now outputs the results to an output file. The user chooses what output file will be created/edited in the terminal window.