Add files via upload
Created a new function in main.cpp called 'void skip_BOM_encoding(std::ifstream& ifs).' If a file uses UTF-8 with BOM (instead of just UTF-8), this function will detect the BOM encoding (the first 3 bytes at the beginning of the input file) and overpass it. Doing this will allow the lexer/syntax_analyzer to read the file in the same way as a UTF-8 file.